# apps/rate-limiter-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Rate Limiter SaaS API
  description: >
    Multi-tenant distributed rate limiter. Tenants sign up, create API keys,
    and configure rate-limit rules (fixed window, sliding window, leaky
    bucket, or token bucket) per resource. Client applications then call
    `/v1/rate-limit/check` with their API key to enforce those limits.
  version: 1.0.0
servers:
  - url: http://localhost:3333
    description: Local development
...
security: []
paths:
...
  /v1/auth/signup:
    post:
      tags: [Auth]
      operationId: signup
      summary: Create a tenant account and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '201':
          description: Account created; session cookie set
          headers:
            Set-Cookie:
              schema: { type: string, example: 'rl_session=<jwt>; Path=/; HttpOnly; SameSite=Lax' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409':
          description: An account with this email already exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /v1/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Log in and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Logged in; session cookie set
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401':
          description: Invalid email or password
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /v1/auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentTenant
      summary: Get the currently signed-in tenant
      security: [{ SessionAuth: [] }]
      responses:
        '200':
          description: The signed-in tenant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ------------------------------------------------------------- Account
  /v1/account/api-keys:
    get:
      tags: [Account]
      operationId: listAccountApiKeys
      summary: List API keys for the signed-in tenant
      security: [{ SessionAuth: [] }]
      responses:
        '200':
          description: The tenant's API keys
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ApiKeySummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Account]
      operationId: createAccountApiKey
      summary: Create a new API key for the signed-in tenant
      security: [{ SessionAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '201':
          description: >
            The new key. `key` (the raw value) is returned only this once —
            only its hash is persisted, so save it now.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatedApiKey' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/account/api-keys/{id}:
    delete:
      tags: [Account]
      operationId: revokeAccountApiKey
      summary: Revoke an API key belonging to the signed-in tenant
      security: [{ SessionAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ApiKeyId'
      responses:
        '204': { description: Revoked }
        '401':
          description: Not signed in, or the key does not belong to this tenant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /v1/tenants:
    post:
      tags: [Admin]
      operationId: createTenant
      summary: Create a tenant (operator bootstrap, no password/login access)
      security: [{ AdminTokenAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email]
              properties:
                name: { type: string }
                email: { type: string, format: email }
      responses:
        '201':
          description: The created tenant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/tenants/{tenantId}/api-keys:
    post:
      tags: [Admin]
      operationId: createTenantApiKey
      summary: Create an API key for a tenant
      security: [{ AdminTokenAuth: [] }]
      parameters:
        - name: tenantId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '201':
          description: >
            The new key. `key` (the raw value) is returned only this once —
            only its hash is persisted, so save it now.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatedApiKey' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/api-keys/{id}:
    delete:
      tags: [Admin]
      operationId: revokeApiKey
      summary: Revoke any API key
      security: [{ AdminTokenAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ApiKeyId'
      responses:
        '204': { description: Revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # --------------------------------------------------------------- Rules
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: A tenant's raw API key (shown once at creation time).
    AdminTokenAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: Static operator token (ADMIN_TOKEN env var).
    SessionAuth:
      type: apiKey
      in: cookie
      name: rl_session
      description: httpOnly JWT session cookie set by /v1/auth/signup or /v1/auth/login.
  parameters:
...
    ApiKeyId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
    Tenant:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        createdAt: { type: string, format: date-time }
    ApiKeyStatus:
      type: string
      enum: [ACTIVE, REVOKED]
    ApiKeySummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        prefix: { type: string, example: rlk_ab12cd34 }
        status: { $ref: '#/components/schemas/ApiKeyStatus' }
        createdAt: { type: string, format: date-time }
        lastUsedAt: { type: string, format: date-time, nullable: true }
    CreatedApiKey:
      allOf:
        - $ref: '#/components/schemas/ApiKeySummary'
        - type: object
          properties:
            key:
              type: string
              description: The raw key. Shown only in this response — save it now.
              example: rlk_ab12cd34EFghijKLmnopQRstuv
...